fix: move create token to browswer
